Update NaryTree.h
Added a trivia array that keeps track of which artist has the most albums, which album has the most tracks, and which track has the most plays. Also added remove functions that take in a string for an artist, album, or track, finds all instances of that string in the appropriate level in the N-ary tree and then removes it from the tree (adjusting the rest of the tree accordingly and printing the results).